requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0 
fastapi==0.116.1
uvicorn==0.35.0

pydantic-settings==2.1.0